pyproject.toml
[project]
name = "diffusion-assessment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch",
    "torchvision",
    "Pillow",
    "tqdm",
    "open-clip-torch",
    "pyyaml",
    "diffusers",
    "transformers",
    "accelerate",
    "safetensors",
    "numpy",
    "scipy",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "pandas>=2.3.2",
]
